Code cleanup
Move Editor classes to editor namespace, better method structure, Rider code cleanup
