+++ Editor/mukaderabbit.humr.Editor.asmdef
@@ -1,16 +1,27 @@
 {
-    "name": "mukaderabbit.humr.Editor",
-    "rootNamespace": "HUMR",
-    "references": [
-        "GUID:4156b7325025f45479ad99b85e7efd7c"
-    ],
-    "includePlatforms": [],
-    "excludePlatforms": [],
-    "allowUnsafeCode": false,
-    "overrideReferences": false,
-    "precompiledReferences": [],
-    "autoReferenced": true,
-    "defineConstraints": [],
-    "versionDefines": [],
-    "noEngineReferences": false
+  "name": "Mukaderabbit.Humr.Editor",
+  "rootNamespace": "",
+  "references": [
+    "GUID:4156b7325025f45479ad99b85e7efd7c",
+    "GUID:84265b35cca3905448e623ef3903f0ff",
+    "GUID:99835874ee819da44948776e0df4ff1d",
+    "GUID:3c1bc1267eab5884ebe7f232c09ee0d9",
+    "GUID:627c4d5cd580ddf41bd320e784fe8b9d",
+    "GUID:1b3abf7e59395d84c84b48d69677daaa",
+    "GUID:6e9c6119ac4eb334284fb7b4bc6d1f05",
+    "GUID:21332e1f0d937794d916d2402ba1943a",
+    "GUID:0f9e8f73c69964f1e8176f9a0b86f9b2",
+    "GUID:f51ebe6a0ceec4240a699833d6309b23"
+  ],
+  "includePlatforms": [
+    "Editor"
+  ],
+  "excludePlatforms": [],
+  "allowUnsafeCode": false,
+  "overrideReferences": false,
+  "precompiledReferences": [],
+  "autoReferenced": true,
+  "defineConstraints": [],
+  "versionDefines": [],
+  "noEngineReferences": false
 }
