+++ README.md
@@ -4,14 +4,15 @@
 
 A complete rewrite of the codebase so new features can be implemented more easily.
 
-- Combine Recorder and OutputLogLoader packages
-- New OutputLog format: semicolons to separate different types
-- Remove duplicated code
+- Combine Recorder and OutputLogLoader `.unitypackage`s
+- New Recording log format: semicolons to separate different types
+- Select player DisplayName from a dropdown
+- Split up code into different methods
 
 ## Installation
 
 > [!WARNING]
-> Remove `Packages/HUMR OutputLogLoader` and `Prefabs`, `ReadMe`, `Scenes` and `Scripts` in `Assets/HUMR` before importing.
+> Remove the old `HUMR OutputLogLoader` package and `Prefabs`, `ReadMe`, `Scenes` and `Scripts` in `Assets/HUMR` before importing.
 
 ### Requirements
 
@@ -20,11 +21,15 @@ A complete rewrite of the codebase so new features can be implemented more easil
 - FBX Exporter Version 4.2.1
 - VRChat 2026.2.2
 
+Download the `.unitypackage` from releases and import it into your VRChat World project.
+
 ## Usage
+
+Duplicate `Humr Sample Scene.unity` from `Packages/Hyakuashi Udon Motion Recorder/Scenes/` into your `Assets`. Build & Test the world to do a recording, or use [the public world](https://vrchat.com/home/launch?worldId=wrld_5962f8a1-bc92-481e-b05a-7cb90eadce34). Use the `PlayerRecordingLoader` component in `Avatar_Utility Loader` to Load recording and export fbx.
 
 ## Changelog
 
-[CHANGELOG.md](CHANGELOG.md#200-beta0---Unreleased)
+[CHANGELOG.md](CHANGELOG.md#200-beta0---2026-07-16)
 
 ## License
