--- Runtime/CSharp/RecordingScanner.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: e378acd2c57845d6bd3f8ebc5486e7d9
-timeCreated: 1782915483
