+++ Runtime/mukaderabbit.humr.Runtime.asmdef
@@ -1,19 +1,19 @@
 {
-    "name": "mukaderabbit.humr.Runtime",
-    "rootNamespace": "HUMR",
-    "references": [
-        "GUID:99835874ee819da44948776e0df4ff1d",
-        "GUID:77fc2e1f5a4f24efab94081ad67fcb20",
-        "GUID:0f9e8f73c69964f1e8176f9a0b86f9b2",
-        "GUID:84265b35cca3905448e623ef3903f0ff"
-    ],
-    "includePlatforms": [],
-    "excludePlatforms": [],
-    "allowUnsafeCode": false,
-    "overrideReferences": false,
-    "precompiledReferences": [],
-    "autoReferenced": true,
-    "defineConstraints": [],
-    "versionDefines": [],
-    "noEngineReferences": false
+  "name": "Mukaderabbit.Humr.Runtime",
+  "rootNamespace": "",
+  "references": [
+    "GUID:99835874ee819da44948776e0df4ff1d",
+    "GUID:3c1bc1267eab5884ebe7f232c09ee0d9",
+    "GUID:f70a94b0bedfa8ec50ed757f72032810",
+    "GUID:75469ad4d38634e559750d17036d5f7c"
+  ],
+  "includePlatforms": [],
+  "excludePlatforms": [],
+  "allowUnsafeCode": false,
+  "overrideReferences": false,
+  "precompiledReferences": [],
+  "autoReferenced": true,
+  "defineConstraints": [],
+  "versionDefines": [],
+  "noEngineReferences": false
 }
