+++ Runtime/mukaderabbit.humr.Runtime.asset
@@ -10,6 +10,6 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 5136146375e9a0a498a72a0091b40cc1, type: 3}
-  m_Name: mukaderabbit.humr.Runtime
+  m_Name: Mukaderabbit.Humr.Runtime
   m_EditorClassIdentifier: 
   sourceAssembly: {fileID: 5897886265953266890, guid: 4156b7325025f45479ad99b85e7efd7c, type: 3}
